[feature/account-controller] небольшой рефакторинг OAuthAccount
